major/major_ws/src/detection/detection/Leaf_Detection_Project/requirements.txt
torch
torchvision
tensorflow
tensorflow-datasets
opencv-python
pillow
scipy
numpy
matplotlib
seaborn
pandas
scikit-learn
tqdm